+++ components/modal/login-modal/constants.js
@@ -1,0 +1,20 @@
+export const FORM_ELEMENTS = {
+  elements: {},
+  validate() {
+    const { elements } = this;
+    Object.keys(elements).forEach((k) => {
+      elements[k].validate();
+    });
+  },
+  isValid() {
+    const { elements } = this;
+    const valid = Object.keys(elements)
+      .map(k => elements[k].isValid())
+      .filter(v => v !== null)
+      .every(element => element);
+
+    return valid;
+  }
+};
+
+export default { FORM_ELEMENTS };
