+++ css/_deprecated.scss
@@ -48,6 +48,7 @@ $fullscreen-margin: 24px;
 $space-1: 8px;
 
 $max-width-modal: 880px;
+$max-width-modal-large: 760px;
 $max-width-modal-medium: 600px;
 
 // Theme [Deprecated]
